Fix Tinc firewall rules
Renamed the tinc configuration template to use the tinc port
defined in tinc pillar. So changing the tinc port will also
be reflected in the firewall configuration.

I've added the tinc ferm rule before bird starts for the reason
bird require the tinc tunnels for connecting to InterCity VPN
peers.